requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argcomplete==1.10.3
asttokens==3.0.0
beautifulsoup4==4.8.2
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
comm==0.2.2
compressed-rtf==1.0.7
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
docx2txt==0.9
ebcdic==1.1.1
executing==2.2.0
extract-msg==0.28.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
olefile==0.47
openai==1.76.2
packaging==25.0
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20191110
pdfplumber==0.11.6
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
PyMuPDF==1.25.5
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-pptx==0.6.23
pyzmq==26.4.0
setuptools==80.1.0
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
stack-data==0.6.3
textract==1.6.5
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzlocal==5.3.1
wcwidth==0.2.13
xlrd==1.2.0
XlsxWriter==3.2.3
